--- src/Chart/Chart.scss
@@ -1,14 +1,0 @@
-#chart-wrapper {
-	margin: 80px 30px;
-	max-width: 1800px;
-	min-width: 1000px;
-	height: 600px;
-	position: relative;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	background: var(--gradient-light);
-	color: black;
-	border-radius: 8px;
-}
